+++ example/lib/main.dart
@@ -1,6 +1,7 @@
 //  Copyright (c) 2019 Aleksander Woźniak
 //  Licensed under Apache License v2.0
 
+import 'package:date_utils/date_utils.dart';
 import 'package:flutter/material.dart';
 import 'package:table_calendar/table_calendar.dart';
 
@@ -28,26 +29,69 @@ class MyHomePage extends StatefulWidget {
   _MyHomePageState createState() => _MyHomePageState();
 }
 
-class _MyHomePageState extends State<MyHomePage> {
+class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
   DateTime _selectedDay;
   Map<DateTime, List> _events;
+  Map<DateTime, List> _visibleEvents;
   List _selectedEvents;
+  AnimationController _controller;
 
   @override
   void initState() {
     super.initState();
-    final now = DateTime.now();
-    _selectedDay = DateTime(now.year, now.month, now.day);
+    _selectedDay = DateTime.now();
     _events = {
-      DateTime(2019, 3, 1): ['Event A', 'Event B', 'Event C'],
-      DateTime(2019, 3, 4): ['Event A'],
-      DateTime(2019, 3, 5): ['Event B', 'Event C'],
-      DateTime(2019, 3, 13): ['Event A', 'Event B', 'Event C'],
-      DateTime(2019, 3, 15): ['Event A', 'Event B', 'Event C', 'Event D', 'Event E', 'Event F', 'Event G'],
-      DateTime(2019, 2, 26): Set.from(['Event A', 'Event A', 'Event B']).toList(),
-      DateTime(2019, 2, 18): ['Event A', 'Event A', 'Event B'],
+      _selectedDay.subtract(Duration(days: 30)): ['Event A0', 'Event B0', 'Event C0'],
+      _selectedDay.subtract(Duration(days: 27)): ['Event A1'],
+      _selectedDay.subtract(Duration(days: 20)): ['Event A2', 'Event B2', 'Event C2', 'Event D2'],
+      _selectedDay.subtract(Duration(days: 16)): ['Event A3', 'Event B3'],
+      _selectedDay.subtract(Duration(days: 10)): ['Event A4', 'Event B4', 'Event C4'],
+      _selectedDay.subtract(Duration(days: 4)): ['Event A5', 'Event B5', 'Event C5'],
+      _selectedDay.subtract(Duration(days: 2)): ['Event A6', 'Event B6'],
+      _selectedDay: ['Event A7', 'Event B7', 'Event C7', 'Event D7'],
+      _selectedDay.add(Duration(days: 1)): ['Event A8', 'Event B8', 'Event C8', 'Event D8', 'Event E8', 'Event F8', 'Event G8'],
+      _selectedDay.add(Duration(days: 3)): Set.from(['Event A9', 'Event A9', 'Event B9']).toList(),
+      _selectedDay.add(Duration(days: 7)): ['Event A10', 'Event B10', 'Event C10'],
+      _selectedDay.add(Duration(days: 11)): ['Event A11', 'Event B11'],
+      _selectedDay.add(Duration(days: 17)): ['Event A12', 'Event B12', 'Event C12', 'Event D12'],
+      _selectedDay.add(Duration(days: 22)): ['Event A13', 'Event B13'],
+      _selectedDay.add(Duration(days: 26)): ['Event A14', 'Event B14', 'Event C14'],
     };
     _selectedEvents = _events[_selectedDay] ?? [];
+    _visibleEvents = _events;
+
+    _controller = AnimationController(
+      vsync: this,
+      duration: const Duration(milliseconds: 400),
+    );
+
+    _controller.forward();
+  }
+
+  void _onDaySelected(DateTime day, List events) {
+    setState(() {
+      _selectedDay = day;
+      _selectedEvents = events;
+    });
+
+    print('Selected day: $day');
+  }
+
+  void _onFormatChanged(CalendarFormat format) {
+    print('Current format: $format');
+  }
+
+  void _onVisibleDaysChanged(DateTime first, DateTime last) {
+    setState(() {
+      _visibleEvents = Map.fromEntries(
+        _events.entries.where(
+          (entry) => entry.key.isAfter(first.subtract(const Duration(days: 1))) && entry.key.isBefore(last.add(const Duration(days: 1))),
+        ),
+      );
+    });
+
+    print('First visible day: $first');
+    print('Last visible day: $last');
   }
 
   @override
@@ -59,7 +103,10 @@ class _MyHomePageState extends State<MyHomePage> {
       body: Column(
         mainAxisSize: MainAxisSize.max,
         children: <Widget>[
+          // Switch out 2 lines below to play with TableCalendar's settings
+          //-----------------------
           _buildTableCalendar(),
+          // _buildTableCalendarWithBuilders(),
           const SizedBox(height: 8.0),
           Expanded(child: _buildEventList()),
         ],
@@ -67,10 +114,10 @@ class _MyHomePageState extends State<MyHomePage> {
     );
   }
 
-  // Configure the calendar here
+  // Simple TableCalendar configuration (using Styles)
   Widget _buildTableCalendar() {
     return TableCalendar(
-      events: _events,
+      events: _visibleEvents,
       initialCalendarFormat: CalendarFormat.week,
       formatAnimation: FormatAnimation.slide,
       startingDayOfWeek: StartingDayOfWeek.monday,
@@ -83,7 +130,7 @@ class _MyHomePageState extends State<MyHomePage> {
       calendarStyle: CalendarStyle(
         selectedColor: Colors.deepOrange[400],
         todayColor: Colors.deepOrange[200],
-        eventMarkerColor: Colors.brown[700],
+        markersColor: Colors.brown[700],
       ),
       headerStyle: HeaderStyle(
         formatButtonTextStyle: TextStyle().copyWith(color: Colors.white, fontSize: 15.0),
@@ -92,12 +139,99 @@ class _MyHomePageState extends State<MyHomePage> {
           borderRadius: BorderRadius.circular(16.0),
         ),
       ),
-      onDaySelected: (day) {
-        setState(() {
-          _selectedDay = day;
-          _selectedEvents = _events[_selectedDay] ?? [];
-        });
+      onDaySelected: _onDaySelected,
+      onFormatChanged: _onFormatChanged,
+      onVisibleDaysChanged: _onVisibleDaysChanged,
+    );
+  }
+
+  // More advanced TableCalendar configuration (using Builders & Styles)
+  Widget _buildTableCalendarWithBuilders() {
+    return TableCalendar(
+      events: _visibleEvents,
+      initialCalendarFormat: CalendarFormat.month,
+      formatAnimation: FormatAnimation.slide,
+      startingDayOfWeek: StartingDayOfWeek.sunday,
+      availableGestures: AvailableGestures.all,
+      availableCalendarFormats: [
+        CalendarFormat.month,
+        CalendarFormat.week,
+      ],
+      calendarStyle: CalendarStyle(
+        outsideDaysVisible: false,
+        weekendStyle: TextStyle().copyWith(color: Colors.blue[800]),
+        outsideWeekendStyle: TextStyle().copyWith(color: Colors.blue[800].withAlpha(127)),
+      ),
+      daysOfWeekStyle: DaysOfWeekStyle(
+        weekendStyle: TextStyle().copyWith(color: Colors.blue[600]),
+      ),
+      headerStyle: HeaderStyle(
+        centerHeaderTitle: true,
+        formatButtonVisible: false,
+      ),
+      builders: CalendarBuilders(
+        selectedDayBuilder: (context, date, _) {
+          return FadeTransition(
+            opacity: Tween(begin: 0.0, end: 1.0).animate(_controller),
+            child: Container(
+              margin: const EdgeInsets.all(4.0),
+              padding: const EdgeInsets.only(top: 5.0, left: 6.0),
+              color: Colors.deepOrange[300],
+              width: 100,
+              height: 100,
+              child: Text(
+                '${date.day}',
+                style: TextStyle().copyWith(fontSize: 16.0),
+              ),
+            ),
+          );
+        },
+        todayDayBuilder: (context, date, _) {
+          return Container(
+            margin: const EdgeInsets.all(4.0),
+            padding: const EdgeInsets.only(top: 5.0, left: 6.0),
+            color: Colors.amber[400],
+            width: 100,
+            height: 100,
+            child: Text(
+              '${date.day}',
+              style: TextStyle().copyWith(fontSize: 16.0),
+            ),
+          );
+        },
+        markersBuilder: (context, date, events) {
+          return Positioned(
+            right: 1,
+            bottom: 1,
+            child: AnimatedContainer(
+              duration: const Duration(milliseconds: 400),
+              decoration: BoxDecoration(
+                shape: BoxShape.rectangle,
+                color: Utils.isSameDay(date, _selectedDay)
+                    ? Colors.brown[400]
+                    : Utils.isSameDay(date, DateTime.now()) ? Colors.brown[300] : Colors.blue[400],
+              ),
+              width: 16.0,
+              height: 16.0,
+              child: Center(
+                child: Text(
+                  '${events.length}',
+                  style: TextStyle().copyWith(
+                    color: Colors.white,
+                    fontSize: 12.0,
+                  ),
+                ),
+              ),
+            ),
+          );
+        },
+      ),
+      onDaySelected: (date, events) {
+        _onDaySelected(date, events);
+        _controller.forward(from: 0.0);
       },
+      onFormatChanged: _onFormatChanged,
+      onVisibleDaysChanged: _onVisibleDaysChanged,
     );
   }