+++ lib/src/widgets/widgets.dart
@@ -1,2 +1,5 @@
+//  Copyright (c) 2019 Aleksander Woźniak
+//  Licensed under Apache License v2.0
+
 export 'cell_widget.dart';
 export 'custom_icon_button.dart';