+++ lib/table_calendar.dart
@@ -3,18 +3,25 @@
 
 library table_calendar;
 
+import 'package:date_utils/date_utils.dart';
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 import 'package:simple_gesture_detector/simple_gesture_detector.dart';
 
+import 'src/customization/customization.dart';
 import 'src/logic/calendar_logic.dart';
-import 'src/styles/styles.dart';
 import 'src/widgets/widgets.dart';
 
-export 'src/styles/styles.dart';
+export 'src/customization/customization.dart';
 
-typedef void _OnDaySelected(DateTime day);
-typedef void _OnFormatChanged(CalendarFormat format);
+/// Callback exposing currently selected day.
+typedef void OnDaySelected(DateTime day, List events);
+
+/// Callback exposing current `CalendarFormat`.
+typedef void OnFormatChanged(CalendarFormat format);
+
+/// Callback exposing currently visible days (first and last of them).
+typedef void OnVisibleDaysChanged(DateTime first, DateTime last);
 
 /// Format to display the `TableCalendar` with.
 enum CalendarFormat { month, twoWeeks, week }
@@ -30,18 +37,20 @@ enum StartingDayOfWeek { monday, sunday }
 /// Gestures available to interal `TableCalendar`'s logic.
 enum AvailableGestures { none, verticalSwipe, horizontalSwipe, all }
 
-/// Highly customizable Calendar widget organized neatly into a `Table`.
-/// Autosizes vertically, saving space for other widgets.
+/// Highly customizable, feature-packed Flutter Calendar with gestures, animations and multiple formats.
 class TableCalendar extends StatefulWidget {
   /// Contains a `List` of objects (eg. events) assigned to particular `DateTime`s.
   /// Each `DateTime` inside this `Map` should get its own `List` of above mentioned objects.
   final Map<DateTime, List> events;
 
   /// Called whenever any day gets tapped.
-  final _OnDaySelected onDaySelected;
+  final OnDaySelected onDaySelected;
 
   /// Called whenever `CalendarFormat` changes.
-  final _OnFormatChanged onFormatChanged;
+  final OnFormatChanged onFormatChanged;
+
+  /// Called whenever the range of visible days changes.
+  final OnVisibleDaysChanged onVisibleDaysChanged;
 
   /// Initially selected DateTime. Usually it will be `DateTime.now()`.
   final DateTime initialDate;
@@ -75,9 +84,15 @@ class TableCalendar extends StatefulWidget {
   /// Use `StartingDayOfWeek.sunday` for Sunday - Saturday week format.
   final StartingDayOfWeek startingDayOfWeek;
 
+  /// `HitTestBehavior` for every day cell inside `TableCalendar`.
+  final HitTestBehavior dayHitTestBehavior;
+
   /// Specify Gestures available to `TableCalendar`.
   /// If `AvailableGestures.none` is used, the Calendar will only be interactive via buttons.
   final AvailableGestures availableGestures;
+
+  /// Configuration for vertical Swipe detector.
+  final SimpleSwipeConfig simpleSwipeConfig;
 
   /// Style for `TableCalendar`'s content.
   final CalendarStyle calendarStyle;
@@ -88,11 +103,15 @@ class TableCalendar extends StatefulWidget {
   /// Style for `TableCalendar`'s Header.
   final HeaderStyle headerStyle;
 
+  /// Set of Builders for `TableCalendar` to work with.
+  final CalendarBuilders builders;
+
   TableCalendar({
     Key key,
     this.events = const {},
     this.onDaySelected,
     this.onFormatChanged,
+    this.onVisibleDaysChanged,
     this.initialDate,
     this.initialCalendarFormat = CalendarFormat.month,
     this.forcedCalendarFormat,
@@ -100,10 +119,13 @@ class TableCalendar extends StatefulWidget {
     this.headerVisible = true,
     this.formatAnimation = FormatAnimation.slide,
     this.startingDayOfWeek = StartingDayOfWeek.sunday,
+    this.dayHitTestBehavior = HitTestBehavior.deferToChild,
     this.availableGestures = AvailableGestures.all,
+    this.simpleSwipeConfig = const SimpleSwipeConfig(verticalThreshold: 20.0, swipeDetectionMoment: SwipeDetectionMoment.onUpdate),
     this.calendarStyle = const CalendarStyle(),
     this.daysOfWeekStyle = const DaysOfWeekStyle(),
     this.headerStyle = const HeaderStyle(),
+    this.builders = const CalendarBuilders(),
   })  : assert(availableCalendarFormats.contains(initialCalendarFormat)),
         assert(availableCalendarFormats.length <= CalendarFormat.values.length),
         super(key: key);
@@ -116,54 +138,53 @@ class TableCalendar extends StatefulWidget {
 
 class _TableCalendarState extends State<TableCalendar> with SingleTickerProviderStateMixin {
   CalendarLogic _calendarLogic;
-  double _dx;
 
   @override
   void initState() {
     super.initState();
     _calendarLogic = CalendarLogic(
-      widget.initialCalendarFormat,
       widget.availableCalendarFormats,
       widget.startingDayOfWeek,
+      initialFormat: widget.initialCalendarFormat,
       initialDate: widget.initialDate,
+      onFormatChanged: widget.onFormatChanged,
+      onVisibleDaysChanged: widget.onVisibleDaysChanged,
     );
-    _dx = 0;
+  }
+
+  @override
+  void dispose() {
+    _calendarLogic.dispose();
+    super.dispose();
   }
 
   void _selectPrevious() {
     setState(() {
       _calendarLogic.selectPrevious();
     });
-
-    _dx = -1.2;
   }
 
   void _selectNext() {
     setState(() {
       _calendarLogic.selectNext();
     });
-
-    _dx = 1.2;
   }
 
   void _selectDate(DateTime date) {
     setState(() {
       _calendarLogic.selectedDate = date;
-    });
 
-    if (widget.onDaySelected != null) {
-      widget.onDaySelected(date);
-    }
+      if (widget.onDaySelected != null) {
+        final key = widget.events.keys.firstWhere((it) => Utils.isSameDay(it, date), orElse: () => null);
+        widget.onDaySelected(date, widget.events[key] ?? []);
+      }
+    });
   }
 
   void _toggleCalendarFormat() {
     setState(() {
       _calendarLogic.toggleCalendarFormat();
     });
-
-    if (widget.onFormatChanged != null) {
-      widget.onFormatChanged(_calendarLogic.calendarFormat);
-    }
   }
 
   void _onHorizontalSwipe(DismissDirection direction) {
@@ -193,9 +214,11 @@ class _TableCalendarState extends State<TableCalendar> with SingleTickerProvider
       const SizedBox(height: 4.0),
     ]);
 
-    return Column(
-      mainAxisSize: MainAxisSize.min,
-      children: children,
+    return ClipRect(
+      child: Column(
+        mainAxisSize: MainAxisSize.min,
+        children: children,
+      ),
     );
   }
 
@@ -320,10 +343,7 @@ class _TableCalendarState extends State<TableCalendar> with SingleTickerProvider
           _calendarLogic.swipeCalendarFormat(false);
         });
       },
-      swipeConfig: SimpleSwipeConfig(
-        verticalThreshold: 30.0,
-        swipeDetectionMoment: SwipeDetectionMoment.onUpdate,
-      ),
+      swipeConfig: widget.simpleSwipeConfig,
     );
   }
 
@@ -333,7 +353,7 @@ class _TableCalendarState extends State<TableCalendar> with SingleTickerProvider
       switchInCurve: Curves.decelerate,
       transitionBuilder: (child, animation) {
         return SlideTransition(
-          position: Tween<Offset>(begin: Offset(_dx, 0), end: Offset(0, 0)).animate(animation),
+          position: Tween<Offset>(begin: Offset(_calendarLogic.dx, 0), end: Offset(0, 0)).animate(animation),
           child: child,
         );
       },
@@ -349,23 +369,15 @@ class _TableCalendarState extends State<TableCalendar> with SingleTickerProvider
   }
 
   Widget _buildTable() {
-    final children = <TableRow>[];
     final daysInWeek = 7;
-    final calendarFormat = widget.forcedCalendarFormat != null ? widget.forcedCalendarFormat : _calendarLogic.calendarFormat;
+    final children = <TableRow>[
+      _buildDaysOfWeek(),
+    ];
 
-    children.add(_buildDaysOfWeek());
-
-    if (calendarFormat == CalendarFormat.week) {
-      children.add(_buildTableRow(_calendarLogic.visibleWeek.toList()));
-    } else if (calendarFormat == CalendarFormat.twoWeeks) {
-      children.add(_buildTableRow(_calendarLogic.visibleTwoWeeks.take(daysInWeek).toList()));
-      children.add(_buildTableRow(_calendarLogic.visibleTwoWeeks.skip(daysInWeek).toList()));
-    } else {
-      int x = 0;
-      while (x < _calendarLogic.visibleMonth.length) {
-        children.add(_buildTableRow(_calendarLogic.visibleMonth.skip(x).take(daysInWeek).toList()));
-        x += daysInWeek;
-      }
+    int x = 0;
+    while (x < _calendarLogic.visibleDays.length) {
+      children.add(_buildTableRow(_calendarLogic.visibleDays.skip(x).take(daysInWeek).toList()));
+      x += daysInWeek;
     }
 
     return Table(
@@ -377,7 +389,7 @@ class _TableCalendarState extends State<TableCalendar> with SingleTickerProvider
 
   TableRow _buildDaysOfWeek() {
     return TableRow(
-      children: _calendarLogic.visibleWeek.map((date) {
+      children: _calendarLogic.visibleDays.take(7).map((date) {
         return Center(
           child: Text(
             DateFormat.E().format(date),
@@ -400,57 +412,116 @@ class _TableCalendarState extends State<TableCalendar> with SingleTickerProvider
               maxHeight: constraints.maxWidth,
               minHeight: constraints.maxWidth,
             ),
-            child: _buildCellContent(date),
+            child: _buildCell(date),
           ),
     );
   }
 
-  Widget _buildCellContent(DateTime date) {
-    Widget content = CellWidget(
-      text: '${date.day}',
-      isSelected: _calendarLogic.isSelected(date),
-      isToday: _calendarLogic.isToday(date),
-      isWeekend: _calendarLogic.isWeekend(date),
-      isOutsideMonth: _calendarLogic.isExtraDay(date),
-      calendarStyle: widget.calendarStyle,
-    );
+  Widget _buildCell(DateTime date) {
+    if (!widget.calendarStyle.outsideDaysVisible &&
+        _calendarLogic.isExtraDay(date) &&
+        _calendarLogic.calendarFormat == CalendarFormat.month) {
+      return Container();
+    }
 
-    if (widget.events.containsKey(date) && widget.events[date].isNotEmpty) {
+    Widget content = _buildCellContent(date);
+
+    final key = widget.events.keys.firstWhere((it) => Utils.isSameDay(it, date), orElse: () => null);
+    if (key != null && widget.events[key].isNotEmpty) {
       final children = <Widget>[content];
-      final maxMarkers = 4;
 
-      children.add(
-        Positioned(
-          bottom: 5.0,
-          child: Row(
-            mainAxisSize: MainAxisSize.min,
-            children: widget.events[date].take(maxMarkers).map((_) => _buildMarker()).toList(),
+      if (widget.builders.markersBuilder != null) {
+        children.add(
+          Builder(
+            builder: (context) => widget.builders.markersBuilder(context, key, widget.events[key]),
           ),
-        ),
-      );
+        );
+      } else {
+        children.add(
+          Positioned(
+            top: widget.calendarStyle.markersPositionTop,
+            bottom: widget.calendarStyle.markersPositionBottom,
+            left: widget.calendarStyle.markersPositionLeft,
+            right: widget.calendarStyle.markersPositionRight,
+            child: Row(
+              mainAxisSize: MainAxisSize.min,
+              children: widget.events[key]
+                  .take(widget.calendarStyle.markersMaxAmount)
+                  .map(
+                    (event) => _buildMarker(key, event),
+                  )
+                  .toList(),
+            ),
+          ),
+        );
+      }
 
       content = Stack(
-        alignment: Alignment.bottomCenter,
+        alignment: widget.calendarStyle.markersAlignment,
         children: children,
       );
     }
 
     return GestureDetector(
-      behavior: HitTestBehavior.opaque,
+      behavior: widget.dayHitTestBehavior,
       onTap: () => _selectDate(date),
       child: content,
     );
   }
 
-  Widget _buildMarker() {
-    return Container(
-      width: 8.0,
-      height: 8.0,
-      margin: const EdgeInsets.symmetric(horizontal: 0.3),
-      decoration: BoxDecoration(
-        shape: BoxShape.circle,
-        color: widget.calendarStyle.eventMarkerColor,
-      ),
-    );
+  Widget _buildCellContent(DateTime date) {
+    final key = widget.events.keys.firstWhere((it) => Utils.isSameDay(it, date), orElse: () => null);
+    if (widget.builders.selectedDayBuilder != null && _calendarLogic.isSelected(date)) {
+      return Builder(
+        builder: (context) => widget.builders.selectedDayBuilder(context, date, widget.events[key]),
+      );
+    } else if (widget.builders.todayDayBuilder != null && _calendarLogic.isToday(date)) {
+      return Builder(
+        builder: (context) => widget.builders.todayDayBuilder(context, date, widget.events[key]),
+      );
+    } else if (widget.builders.outsideWeekendDayBuilder != null && _calendarLogic.isExtraDay(date) && _calendarLogic.isWeekend(date)) {
+      return Builder(
+        builder: (context) => widget.builders.outsideWeekendDayBuilder(context, date, widget.events[key]),
+      );
+    } else if (widget.builders.outsideDayBuilder != null && _calendarLogic.isExtraDay(date)) {
+      return Builder(
+        builder: (context) => widget.builders.outsideDayBuilder(context, date, widget.events[key]),
+      );
+    } else if (widget.builders.weekendDayBuilder != null && _calendarLogic.isWeekend(date)) {
+      return Builder(
+        builder: (context) => widget.builders.weekendDayBuilder(context, date, widget.events[key]),
+      );
+    } else if (widget.builders.dayBuilder != null) {
+      return Builder(
+        builder: (context) => widget.builders.dayBuilder(context, date, widget.events[key]),
+      );
+    } else {
+      return CellWidget(
+        text: '${date.day}',
+        isSelected: _calendarLogic.isSelected(date),
+        isToday: _calendarLogic.isToday(date),
+        isWeekend: _calendarLogic.isWeekend(date),
+        isOutsideMonth: _calendarLogic.isExtraDay(date),
+        calendarStyle: widget.calendarStyle,
+      );
+    }
+  }
+
+  Widget _buildMarker(DateTime date, dynamic event) {
+    if (widget.builders.singleMarkerBuilder != null) {
+      return Builder(
+        builder: (context) => widget.builders.singleMarkerBuilder(context, date, event),
+      );
+    } else {
+      return Container(
+        width: 8.0,
+        height: 8.0,
+        margin: const EdgeInsets.symmetric(horizontal: 0.3),
+        decoration: BoxDecoration(
+          shape: BoxShape.circle,
+          color: widget.calendarStyle.markersColor,
+        ),
+      );
+    }
   }
 }